Made route demand a double value for more flexibility

+++ src/heuristic/popvnd/neighborhood/SwapStations_V2.java
@@ -43,7 +43,7 @@ public final class SwapStations_V2 extends INeighborhood_V2<DemandRoutesSolution
 						
 						int s1 = originalRoute1.get(cs);
 						int s2 = originalRoute2.get(ns);
-						int[] demands = description.demand;
+						double[] demands = description.demand;
 						
 						if(originalRoute1.getDemand()-demands[s1]+demands[s2]>description.capacity ||
 								originalRoute2.getDemand()-demands[s2]+demands[s1]>description.capacity)
